+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/extensions/kotlin/android/application/_internal/configureAndroidApplication.kt
@@ -15,6 +15,7 @@ import com.javiersc.hubdle.extensions.dependencies._internal.constants.ORG_JETBR
 import com.javiersc.hubdle.extensions.dependencies._internal.constants.ORG_JETBRAINS_KOTLINX_KOTLINX_SERIALIZATION_JSON_MODULE
 import com.javiersc.hubdle.extensions.dependencies._internal.constants.ORG_JETBRAINS_KOTLIN_KOTLIN_TEST_MODULE
 import com.javiersc.hubdle.extensions.kotlin._internal.configJvmTarget
+import com.javiersc.hubdle.extensions.kotlin.android._internal.configureAndroidBuildFeatures
 import com.javiersc.hubdle.extensions.options.configDefaultAndroidSourceSets
 import com.javiersc.hubdle.extensions.options.configureDefaultKotlinSourceSets
 import org.gradle.api.Project
@@ -44,6 +45,7 @@ internal fun configureAndroidApplication(project: Project) {
             // TODO: namespace = project.calculateAndroidNamespace(androidState.namespace)
 
             sourceSets.all { configDefaultAndroidSourceSets() }
+            project.configureAndroidBuildFeatures(this)
         }
     }
 }